disk: composite: switch tests to tempfile
Use tempfile rather than SharedMemory for temporary files in the
composite disk unit tests.  This fixes an API incompatibility that was
missed because these tests are only enabled with the composite-disk
feature.

This also enables the disk tests in test_all with the privileged option
so that they run in the VM; this allows the memfd_create syscall used in
the async_read/write tests to work.

BUG=b:181674168
TEST=cargo test -p disk --features=composite-disk
TEST=./test_all

Change-Id: I869b5daa29904b3558440e7b650713e8e5085d81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/2743401
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>